requirements.txt
tkscrolledframe
Pillow
opencv-python
numpy
mediapipe